# bzn-exemplo-services-profiles/src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  title: Swagger Exemplo Services Profiles - OpenAPI 3.0
  description: |-
    Esse é um exemplo de descrição da especificação OpenAPI 3.0 adotada no sistema Brazona Tech.
    Para mais informações consulte o Wiki do Sistema:
    - [The Exemplo Services Perfil repository](https://github.com/brazona/bzn-exemplo-services/tree/main/bzn-exemplo-services-profiles)
    - [The source API definition for the Exemplo Services Perfil](https://github.com/brazona/bzn-exemplo-services/blob/main/bzn-exemplo-services-profiles/src/main/resources/openapi.yaml)
  termsOfService: https://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  version: 1.0.12
servers:
  - url: http://localhost:8484/bzn-exemplo-services-profiles/v1
tags:
  - name: Perfil
    description: API que fornece funcionalidades dos perfils de usuários do sistema.
    externalDocs:
      description: Exemplo Services Profiles Wiki
      url: https://swagger.io
paths:
  /profile/v1/:
    post:
      tags:
        - Perfil
      summary: Adiciona um novo perfil.
      description: Adiciona um novo perfil de usuário.
      operationId: addProfile
      requestBody:
        description: Cria um novo perfil de usuário.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
        default:
          description: Unexpected error
  /profile/v1/all:
    get:
      tags:
        - Perfil
      summary: Busca todos os perfis.
      description: Busca todos os perfils de usuário.
      operationId: getProfileAll
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Profile'
        '400':
          description: Invalid status value
        default:
          description: Unexpected error          

  /profile/v1/{profileID}:
    get:
      tags:
        - Perfil
      summary: Busca perfil pelo ID.
      description: Busca o perfil do usuário pelo ID.
      operationId: getProfileId
      parameters:
        - name: profileID
          in: path
          description: ID do perfil a ser buscado.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Invalid status value
        default:
          description: Unexpected error
    put:
      tags:
        - Perfil
      summary: Atualiza o perfil do usuário.
      description: Atualiza o perfil do usuário pelo ID.
      operationId: updateProfileId
      parameters:
        - name: profileID
          in: path
          description: ID do perfil a ser atualizado.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Atualiza o perfil do usuário pelo ID.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Invalid ID supplied
        '404':
          description: Perfil not found
        '422':
          description: Validation exception
        default:
          description: Unexpected error      
    delete:
      tags:
        - Perfil
      summary: Exclui perfil pelo ID.
      description: Exclui o perfil do usuário pelo ID.
      operationId: deleteProfileId
      parameters:
        - name: profileID
          in: path
          description: ID do perfil a ser excluido.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Perfil deleted
        '400':
          description: Invalid Perfil value
        default:
          description: Unexpected error

components:
  schemas:
    Profile:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identificador do perfil.
          example: 10
        userId:
          type: integer
          format: int64
          description: Identificador do usuário.
          example: 198772
        roleId:
          type: integer
          format: int64
          description: Identificador do papel.
          example: 7
        name:
          type: string
          description: Nome do Perfil
          example: Coodernador
    
    